+++ src/__tests__/pages/about-us/who-we-are.test.tsx
@@ -6,6 +6,7 @@ import AboutWhoWeAre, {
 } from "../../../pages/about-us/who-we-are";
 import CMSClient, { AboutPage } from "../../../node-lib/cms";
 import { mockSeo, portableTextFromString } from "../../__helpers__/cms";
+import renderWithSeo from "../../__helpers__/renderWithSeo";
 
 export const testAboutPageData: AboutPage = {
   title: "About Oak",
@@ -390,21 +391,29 @@ jest.mock("../../../node-lib/cms");
 
 const mockCMSClient = CMSClient as jest.MockedObject<typeof CMSClient>;
 
-describe("pages/about us who we are.tsx", () => {
+describe("pages/about/who-we-are.tsx", () => {
   beforeEach(() => {
     jest.clearAllMocks();
     jest.resetModules();
   });
 
   it("Renders correct title ", async () => {
-    renderWithProviders(
-      <AboutWhoWeAre pageData={testAboutPageData} isPreviewMode={false} />
-    );
+    renderWithProviders(<AboutWhoWeAre pageData={testAboutPageData} />);
 
     await waitFor(() => {
       expect(screen.getByRole("heading", { level: 1 }).textContent).toBe(
         "About us"
       );
+    });
+  });
+
+  describe.skip("SEO", () => {
+    it("renders the correct SEO details", async () => {
+      const { seo } = renderWithSeo(
+        <AboutWhoWeAre pageData={testAboutPageData} />
+      );
+
+      expect(seo).toEqual({});
     });
   });