+++ src/components/VideoPlayer/getDuration.ts
@@ -1,0 +1,17 @@
+import { RefObject } from "react";
+import MuxPlayerElement from "@mux/mux-player";
+
+/**
+ * Gets duration of video from the video player ref.
+ * Rounds down, and returns null if duration not available.
+ */
+const getDuration = (ref: RefObject<MuxPlayerElement>) => {
+  const duration = ref.current?.duration;
+  if (typeof duration === "number" && !isNaN(duration)) {
+    return Math.floor(duration);
+  }
+
+  return null;
+};
+
+export default getDuration;