+++ src/components/pages/Home/HomeAboutCard.tsx
@@ -1,29 +1,42 @@
 import { FC } from "react";
+import { PortableText } from "@portabletext/react";
 
 import useAnalytics from "../../../context/Analytics/useAnalytics";
 import ButtonAsLink from "../../Button/ButtonAsLink";
+import type { Card as CardShape } from "../../../node-lib/cms";
 import Card from "../../Card";
 import CardTitle from "../../Card/CardComponents/CardTitle";
-import { P } from "../../Typography";
+import Typography from "../../Typography";
+import { getCTAHref } from "../../../utils/portableText/resolveInternalHref";
 
-const HomeAboutCard: FC = () => {
+type HomeSidebarTextCard = CardShape;
+
+const HomeAboutCard: FC<HomeSidebarTextCard> = (props) => {
   const { track } = useAnalytics();
 
   return (
     <Card $borderRadius={0} $background="white">
       <CardTitle icon="Search" iconPosition="leading" iconSize={32} tag="h2">
-        About us
+        {props.title}
       </CardTitle>
-      <P color={"black"} fontSize={16} $mb={24}>
-        Discover who we are, what we do and how we work.
-      </P>
-      <ButtonAsLink
-        $mt={"auto"}
-        fullWidth
-        href="/about-us/who-we-are"
-        label="Find out more"
-        htmlAnchorProps={{ onClick: track.aboutSelected }}
-      />
+
+      <Typography color={"black"} $fontSize={16} $mb={24}>
+        <PortableText value={props.bodyPortableText} />
+      </Typography>
+
+      {props.cta && (
+        <ButtonAsLink
+          $mt={"auto"}
+          fullWidth
+          href={getCTAHref(props.cta)}
+          label={props.cta.label}
+          // @TODO: This link is dynamic, not always an about link
+          // so tracking may become incorrect
+          // See owa issue #619
+          // When this is standardized remove the exclusions from sonar
+          htmlAnchorProps={{ onClick: track.aboutSelected }}
+        />
+      )}
     </Card>
   );
 };