+++ src/pages/about-us/board.tsx
@@ -2,7 +2,6 @@ import { NextPage, GetStaticProps } from "next";
 import { PortableText } from "@portabletext/react";
 
 import CMSClient, { AboutPage } from "../../node-lib/cms";
-import { DEFAULT_SEO_PROPS } from "../../browser-lib/seo/Seo";
 import Layout from "../../components/Layout";
 import MaxWidth from "../../components/MaxWidth/MaxWidth";
 import SummaryCard from "../../components/Card/SummaryCard";
@@ -17,22 +16,16 @@ import { reducedAboutNavLinks } from "../../browser-lib/fixtures/aboutNav";
 import AboutIntroCard from "../../components/AboutIntoCard/AboutIntroCard";
 import IconButtonAsLink from "../../components/Button/IconButtonAsLink";
 import Box from "../../components/Box";
+import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
 
 export type AboutPageProps = {
   pageData: AboutPage;
   isPreviewMode: boolean;
 };
 
-const AboutUsBoard: NextPage<AboutPageProps> = ({
-  pageData,
-  isPreviewMode,
-}) => {
+const AboutUsBoard: NextPage<AboutPageProps> = ({ pageData }) => {
   return (
-    <Layout
-      seoProps={DEFAULT_SEO_PROPS}
-      $background={"white"}
-      isPreviewMode={isPreviewMode}
-    >
+    <Layout seoProps={getSeoProps(pageData.seo)} $background={"white"}>
       <MaxWidth $pt={[64, 80]}>
         <SummaryCard
           title={"About us"}