+++ src/pages/lp/[landingPageSlug].tsx
@@ -1,6 +1,6 @@
 import { GetStaticPaths, GetStaticProps, NextPage } from "next";
 
-import { DEFAULT_SEO_PROPS } from "../../browser-lib/seo/Seo";
+import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
 import Grid, { GridArea } from "../../components/Grid";
 import Layout from "../../components/Layout";
 import MaxWidth from "../../components/MaxWidth/MaxWidth";
@@ -11,9 +11,9 @@ export type LandingPageProps = {
   isPreviewMode: boolean;
 };
 
-const Landing: NextPage<LandingPageProps> = ({ pageData, isPreviewMode }) => {
+const Landing: NextPage<LandingPageProps> = ({ pageData }) => {
   return (
-    <Layout seoProps={DEFAULT_SEO_PROPS} isPreviewMode={isPreviewMode}>
+    <Layout seoProps={getSeoProps(pageData.seo)}>
       <MaxWidth>
         <Grid>
           <GridArea $colSpan={[12, 12, 12]}>